% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{washington_roads}
\alias{washington_roads}
\title{World Health Organization TB data}
\format{
## `washinton_roads`
A data frame with 1,501 rows and 9 columns:
\describe{
  \item{ID}{Anonimized road ID. Factor.}
  \item{Year}{Year. Integer.}
  \item{AADT}{Annual Average Daily Traffic (AADT). Double.}
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
washington_roads
}
\description{
Washington road data from Washington Department of Transportation through HSIS.
}
\keyword{datasets}
